+++ src/components/buffer-summary/BufferSummaryVirtualizedList.tsx
@@ -18,8 +18,7 @@ import useBufferNavigation from '../../hooks/useBufferNavigation';
 import BufferSummaryPlotControls from './BufferSummaryPlotControls';
 import { TensorDeallocationReport, TensorsByOperationByAddress } from '../../model/BufferSummary';
 import { CHART_DATA, OPERATION_EL_HEIGHT, TOTAL_SHADE_HEIGHT } from '../../definitions/BufferSummary';
-import { TOP_N_MODE_LABEL, TopNAnnotationMode } from '../../definitions/TopNAnnotations';
-import { RankedAnnotation } from '../../functions/topNAnnotations';
+import { RankedAnnotation, TOP_N_MODE_LABEL, TopNAnnotationMode } from '../../definitions/TopNAnnotations';
 import { perfColorScale } from '../../functions/perfOverlay';
 
 interface BufferSummaryVirtualizedListProps {
