+++ src/functions/sortAndFilterPerfTableData.ts
@@ -5,7 +5,7 @@
 import { ColumnKeys, PerfTableFilters, TypedPerfTableRow, signpostRowDefaults } from '../definitions/PerfTable';
 import { DeviceOperationLayoutTypes } from '../model/APIData';
 import { BufferType } from '../model/BufferType';
-import { Signpost } from '../definitions/PerfSignpost';
+import { Signpost } from '../model/Signpost';
 
 const SIGNPOST_MARKER = '(signpost)';
