+++ src/hooks/useTopNAnnotations.tsx
@@ -10,8 +10,8 @@ import {
     topNAnnotationEnabledAtom,
     topNAnnotationModeAtom,
 } from '../store/app';
-import { TopNAnnotationMode, TopNAnnotationStatus } from '../definitions/TopNAnnotations';
-import { RankedAnnotation, isPerfMode, selectTopNAnnotations } from '../functions/topNAnnotations';
+import { RankedAnnotation, TopNAnnotationMode, TopNAnnotationStatus } from '../definitions/TopNAnnotations';
+import { isPerfMode, selectTopNAnnotations } from '../functions/topNAnnotations';
 import { OpPerfAggregate, PerfOverlaySource, aggregatePerfByOp } from '../functions/perfOverlay';
 import { L1PressureMetrics, L1PressureStatus } from '../model/L1Pressure';
 import { useGetDeviceOperationListPerf, useL1PressureByOperation, usePerformanceReport } from './useAPI';
