+++ src/model/Signpost.ts
