+++ src/store/app.ts
@@ -8,10 +8,11 @@ import { NumberRange, TabId } from '@blueprintjs/core';
 import { Id } from 'react-toastify';
 import { TAB_IDS } from '../definitions/BufferSummary';
 import { ListStates } from '../definitions/VirtualLists';
-import { Signpost } from '../definitions/PerfSignpost';
+import { Signpost } from '../model/Signpost';
 import { PerfTabIds } from '../definitions/Performance';
 import { ReportFolder, ReportLocation } from '../definitions/Reports';
-import { REPORT_LINKS_STORAGE_KEY, ReportLink } from '../definitions/ReportLinks';
+import { REPORT_LINKS_STORAGE_KEY } from '../definitions/ReportLinks';
+import { ReportLink } from '../model/ReportLinks';
 import { ColumnKeys, TypedPerfTableRow } from '../definitions/PerfTable';
 import { BufferType } from '../model/BufferType';
 import { StackedGroupBy } from '../definitions/StackedPerfTable';
