+++ tests/BufferSummaryVirtualizedList.spec.tsx
@@ -10,8 +10,7 @@ import { BufferType } from '../src/model/BufferType';
 import BufferSummaryVirtualizedList from '../src/components/buffer-summary/BufferSummaryVirtualizedList';
 import { ScrollLocations } from '../src/definitions/VirtualLists';
 import { BufferSummaryAxisConfiguration } from '../src/definitions/PlotConfigurations';
-import { TopNAnnotationMode } from '../src/definitions/TopNAnnotations';
-import { RankedAnnotation } from '../src/functions/topNAnnotations';
+import { RankedAnnotation, TopNAnnotationMode } from '../src/definitions/TopNAnnotations';
 
 const memoryPlotRendererMock = vi.fn();
 const bufferSummaryRowMock = vi.fn();
